requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
fastapi==0.109.2
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
numpy==1.26.4
packaging==25.0
pandas==2.2.0
pluggy==1.5.0
pydantic==2.6.1
pydantic_core==2.16.2
pytest==8.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.27.1
